navbar refactored to use grid. window.py also refactored to use grid. cleanup.
