--- View/Input/checkbox.py
@@ -1,26 +1,0 @@
-"""
-checkbox input widget. stores a boolean value that is displaed via a
-checked/not checked box.
-"""
-
-'''
-init(name: str, flag=false: bool): void
-sets default value to false unless otherwise specified
-'''
-
-'''
-do(): void
-class level implementation of this method. emits proper signal to controller.
-also calls swap().
-'''
-
-'''
-click(): void
-event function bound to checkbox click event.
-'''
-
-'''
-swap(): void
-toggles the flag data member. true->false or false->true
-'''
-
