--- View/Input/radio.py
@@ -1,20 +1,0 @@
-"""
-radio button widget. bound to a raidogroup parent. is essentially a 'bucket'
-for a boolean value.
-"""
-
-'''
-init(text: str, selected: bool=false): void
-call super(). set the text and initial value of the radio widget. 
-'''
-
-'''
-click(): void
-click event handler. essentially toggles the truth value and updates the
-View appropriately.
-'''
-
-'''
-do(): void
-preferred binding for all logic.
-'''
