--- View/Input/text.py
@@ -1,19 +1,0 @@
-"""
-root class for all text handling widgets. holds the text string and has
-methods that facilitate data access for all methods.
-"""
-
-'''
-init(text: str): void
-call super(). set default text if needed
-'''
-
-'''
-get_text(): str
-return the 'text' data member as a str.
-'''
-
-'''
-validate(): int
-run input validation. return 0 if good, 1 if bad.
-'''
