--- View/Input/textarea.py
@@ -1,9 +1,0 @@
-"""
-multiline text area. NOT a rich text editor. auto wrap as text gets longer than
-bounds
-"""
-
-'''
-init(text: str, rows: int, columns: int): void
-call super(). create the text area, set placeholder text if needed.
-'''
