--- View/Input/textfield.py
@@ -1,9 +1,0 @@
-"""
-single line text field. has a configurable height and width.
-"""
-
-'''
-init(text: str, width: int, height: int): void
-call super(). initialize any placeholder text, set the width and height of the
-text field.
-'''
