--- View/Layout/gridview.py
@@ -1,12 +1,0 @@
-"""
-wrapper for the tkinter grid layout system that allows for simpler definiton
-of how the elements in a grid should work. takes in a list of children
-and specification of a row an column number.
-"""
-
-'''
-init(name: str, children[]: List<varies>, rows: int, cols: int): void
-call super(). populate the frame using the row and column specification. 
-padding will be set to a default width for now.
-'''
-
