--- View/Layout/page.py
@@ -1,11 +1,0 @@
-"""
-'lowest level' subclass layout widget that will be used more than once. is
-a root container for a specific section of content for the application. 'page'
-seemed an appropriate name.
-"""
-
-'''
-init(name: str, children[]: List<varied>): void
-call super(). as of right now, this widget needs to ensure its height and
-width are set equal to the parent window.
-'''
