--- View/Layout/view.py
@@ -1,8 +1,0 @@
-"""
-generic View/frame class. general purpose in nature.
-"""
-
-'''
-init(name: str, children[]: List<varied>): void
-call super(). populate View with child elements.
-'''
