--- View/Layout/vstack.py
@@ -1,11 +1,0 @@
-"""
-the 'column' class. lays out its children vertically in a View. will wrap
-either left or right if it is not configured to truncate. scrolling is also
-an option.
-"""
-
-'''
-init(name: str, children[]: List<varies>): void
-calls super(). populate vertical View. needs to setup behavior for behaviour 
-that exceeds the bounds of the View.
-'''
