--- View/Widgets/dropdown.py
@@ -1,10 +1,0 @@
-"""
-widget that 'floats' can can be placed anywhere in relation to another widget
-with contents
-"""
-
-'''
-init(name: str, children: List<varied>): void
-call super(). sets the position of the widget and creates the binding to the
-offset widget. initial visible state is false. children widgets are defined.
-'''
