--- View/Widgets/modal.py
@@ -1,11 +1,0 @@
-"""
-widget that creates a modal effect, preventing the user from taking any
-other action until the modal View has been dealt with.
-"""
-
-'''
-init(name: str, children: List<varied>): void
-call super(). needs to set up deactivation on visible. also needs to set up
-proper View release on completion. message text is set
-'''
-
