--- View/Widgets/table.py
@@ -1,10 +1,0 @@
-"""
-parent class for table views contains children rows and columns
-"""
-
-'''
-init(name: str, children: List<varied>): void
-call super(). needs to initialize the proper rows and columns. need to create
-the labels as well as the table header data. have configurable handles that
-allow for filters, sortinc, etc that are bound here.
-'''
