--- View/Widgets/tablecolumn.py
@@ -1,10 +1,0 @@
-"""
-column class for table views. is essentally a series of indices to rows and the
-headers for each column.
-"""
-
-'''
-init(name: str, children: List<varied>): void
-call super(). needs to initialize and create headers. also needs to generate
-the proper indices to each row.
-'''
