--- View/Widgets/tablerow.py
@@ -1,10 +1,0 @@
-"""
-table row View that is a part of a table. has indices to appropriate columns
-and cells can have other widgets embedded, though it will usually be text
-"""
-
-'''
-init(name: str, children: List<varied>): void
-call super(). needs to create an hstack here that is used as the row. also need
-to generate indices to each column.
-'''
