--- View/Widgets/tooltip.py
@@ -1,10 +1,0 @@
-"""
-on hover/on click tooltips for the user. can be configured to show a message
-or even link to external documentation.
-"""
-
-'''
-init(name: str, children: List<varied>): void
-call super(). set tooltip message, bind onhover and onclick events
-'''
-
