--- View/Widgets/widget.py
@@ -1,19 +1,0 @@
-"""
-base class for specialized widgets.
-"""
-
-'''
-init(name: str, children: List<varied>): void
-create root Widget class and set required data. manage visiblility of this and
-all children widgets.
-'''
-
-'''
-get_children(): List<varied>
-DFT to build a list of all children, return the list.
-'''
-
-'''
-toggle(): void
-toggle the visiblilty state of this widget and all child widgets.
-'''
